ppr-ontology/api/requirements-api.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
pyyaml
rdflib>=7.0.0
pyshacl>=0.27.0
httpx>=0.27
# Generation pipeline dependencies
openai>=1.40.0
google-genai>=0.7.0
basyx-python-sdk>=1.2.1,<2.0.0
pymupdf4llm>=0.0.17
